apps/ml-service/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# LLM SDKs
openai==1.10.0

# ML/AI libraries
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
pillow==10.2.0
open-clip-torch==2.24.0

# Database clients
psycopg2-binary==2.9.9
redis==5.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
numpy==1.26.3

# Logging and monitoring
loguru==0.7.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
ruff==0.1.11
